Add tests for validator
